# backend/openapi.yaml
openapi: 3.0.3
info:
  title: POS-B API
  version: 1.0.0
  description: OpenAPI 3 spec pour l'API POS-B (endpoints définis dans routes/api.php)
servers:
  - url: http://localhost
    description: Serveur local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/products:
    get:
      tags: [Products]
      security:
        - bearerAuth: []
    post:
      tags: [Products]
      security:
        - bearerAuth: []
  /api/products/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Products]
      security:
        - bearerAuth: []
    put:
      tags: [Products]
      security:
        - bearerAuth: []
    delete:
      tags: [Products]
      security:
        - bearerAuth: []
...